Doctor Specialist class added
