+++ src/main/java/in/nethaji/service/DoctorService.java
@@ -12,21 +12,7 @@ public class DoctorService {
 		// Default Constructor
 	}
 
-	private static final List<String> specialistList = new ArrayList<>();
-
-	static {
-		specialistList.add("ENT");
-		specialistList.add("General");
-		specialistList.add("Cardiology");
-		specialistList.add("ophthalmologist");
-	}
-
-	public static List<String> getSpecialistList() {
-		return specialistList;
-	}
-
 	private static final List<Doctor> doctorList = new ArrayList<>();
-
 
 	/**
 	 * This method is used to get the list of doctors
@@ -36,7 +22,7 @@ public class DoctorService {
 	public static List<Doctor> getDoctors() {
 		return doctorList;
 	}
-	
+
 	/**
 	 * This method is used to add Doctor details
 	 *
